requirements.txt
absl-py==2.1.0
altair==5.5.0
altgraph==0.17.4
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.6.2
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.11
contourpy==1.3.1
crispy-bootstrap4==2024.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.2.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filter==24.2
django-widget-tweaks==1.4.8
djangorestframework==3.15.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.15.4
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.14.0
holidays==0.70
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
icons==1.1
idna==3.7
imageio==2.37.0
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
Markdown==3.6
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
narwhals==1.24.1
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
optional-django==0.3.0
ortools==9.12.4544
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.2.2
plotly==6.0.0
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
prophet==1.1.6
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyMatting==1.1.14
pyparsing==3.2.1
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pyttsx3==2.98
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
react==4.3.0
referencing==0.35.1
rembg==2.0.67
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
streamlit==1.44.1
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.6.11
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
xgboost==3.0.2
